cms卸载bug

+++ application/cms/uninstall/uninstall.php
@@ -56,8 +56,10 @@ class Uninstall extends UninstallBase
                     }
                 }
             }
-            //会员投稿删除
-            Db::name('member_content')->delete(true);
+            if (isModuleInstall('member')) {
+                //会员投稿删除
+                Db::name('member_content')->delete(true);
+            }
         }
         //删除路由
         if (file_exists(ROOT_PATH . 'route' . DIRECTORY_SEPARATOR . 'route_cms.php')) {
